2/21/14
Worked more on combat, but minus a few compiling error, it should be
fully completed now. But, the package is now acting up, I think it may
be because I have symbols in the folder name, but fi i change it I
can't save the files anymore and I can't commit to GitHub because the
folder doesn't exist. I was playing around with it trying to fix it,
but then I had to revert everything back to commit… super fun.